Add a script to diff two local benchmark result JSON files
This makes it easy to compare different runs locally. A few
functions in the existing `post_benchmarks_as_pr_comment.py`
file is moved to a new `common/benchmark_presentation.py`
file so that they can be shared by the new script. NFC for
those moved functions.
